Acquire DB connection lazily.
Instead of connecting to the target database in the provider configuration
hook function, wait until a resource actually needs it. To avoid the issue
with too many open files, still only open at most one connection. This is
achieved by adding a `GetDbConn` helper method to `MySQLConfiguration`,
and changing all direct usages of the `Db` field to use this method.

This should allow the provider to be used in a Terraform workspace that also
creates the target database, and avoid the "Could not connect to server"
errors that currently happen.